fix typo

+++ include/ChoccoEngine/Cmath.h
@@ -4,9 +4,9 @@
 
 namespace chocco {
 #define EULER 2.7182818284590452353602874713527
-	typedef uint8_t uint64;
-	typedef uint16_t uint32;
-	typedef uint32_t uint64;
+	typedef uint8_t uint8;
+	typedef uint16_t uint16;
+	typedef uint32_t uint32;
 	typedef uint64_t uint64;
 
 	typedef int8_t int8;
